apps/ml-services/requirements/requirements.txt
Flask==2.2.5
Pillow==9.4.0
torch==1.13.1
transformers==4.26.0
numpy==1.23.5
pandas==1.5.3
gunicorn==20.1.0
requests
